backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.42.0
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv==1.0.1
pdfplumber==0.11.0
